start fixing animations
