+++ src/main/main.module.scss
@@ -1,9 +1,8 @@
-.center {
-  position: fixed;
-  width: 50px;
-  height: 50px;
-  left: calc(50vw - 25px);
-  top: calc(50vh - 25px);
-  display: block;
-  background-color: red;
+.planet-wrapper {
+  opacity: 0;
+  display: flex;
+  flex-direction: column;
+  height: 100%;
+  width: 100%;
+  align-items: center;
 }
